feat: made a separate state for the list of categories
