+++ src/store/exports.js
@@ -4,3 +4,4 @@ export { default as modals } from './modals';
 export { default as article } from './article';
 export { default as locale } from './locale';
 export { default as user } from './user'
+export { default as category } from './category'
